+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-07-13 20:20:16
-// Commit: feat(core): add bi-directional event bus handler for core
+// Updated at: 2026-07-15 13:32:24
+// Commit: test(core): add unit tests for core business logic
 
 class coreComponentState {
-  final String commitRef = '2810';
-  final String lastMessage = 'feat(core): add bi-directional event bus handler for core';
-  final DateTime timestamp = DateTime.parse('2026-07-13T20:20:16+05:00');
+  final String commitRef = '2849';
+  final String lastMessage = 'test(core): add unit tests for core business logic';
+  final DateTime timestamp = DateTime.parse('2026-07-15T13:32:24+05:00');
 }